Move hash library to hash package
