+++ capsules/hashit/hash/hash.go
